Internal Feature: Removed unused tasks. [skip ci]
