--- packages/ckeditor5-dev-env/lib/tasks/create-package/templates/.gitignore_template
@@ -1,10 +1,0 @@
-# These files will be ignored by Git and by our linting tools:
-#	grunt jshint
-#	grunt jscs
-#
-# Be sure to append /** to folders to have everything inside them ignored.
-
-# All "dot directories".
-.*/**
-
-node_modules/**
