--- packages/ckeditor5-dev-env/lib/tasks/create-package/templates/.jshintrc
@@ -1,11 +1,0 @@
-{
-  "esnext": true,
-  "immed": true,
-  "loopfunc": true,
-  "noarg": true,
-  "nonbsp": true,
-  "undef": true,
-  "unused": true,
-  "strict": "implied",
-  "varstmt": true
-}
