--- packages/ckeditor5-dev-env/lib/tasks/create-package/templates/CHANGES.md
@@ -1,25 +1,0 @@
-{{AppName}} Changelog
-========================================
-
-## {{AppName}} 0.0.2
-
-**Major|Minor|Patch Release** - Build Date: yyyy-mm-dd
-
-New Features:
-
-* [#Issue Number](http://issue/url): Item 1.
-* Item 2
-
-Fixed Issues:
-
-* [#Issue Number](http://issue/url): Item 1.
-* Item 2
-
-Other Changes:
-
-* Item 1
-* Item 2
-
-## {{AppName}} 0.0.1
-
-...
