--- packages/ckeditor5-dev-env/lib/tasks/create-package/templates/CONTRIBUTING.md
@@ -1,4 +1,0 @@
-Contributing
-========================================
-
-Information about contributing can be found on the following page: <https://github.com/ckeditor/ckeditor5/blob/master/CONTRIBUTING.md>.
