--- packages/ckeditor5-dev-env/lib/tasks/create-package/templates/README.md
@@ -1,10 +1,0 @@
-{{AppName}}
-========================================
-
-[![devDependency Status](https://david-dm.org/{{GitHubRepositoryPath}}/dev-status.svg)](https://david-dm.org/{{GitHubRepositoryPath}}?type=dev)
-
-{{ProjectDescription}} More information about the project can be found at the following URL: <https://github.com/{{GitHubRepositoryPath}}>.
-
-## License
-
-Licensed under the GPL, LGPL and MPL licenses, at your choice. For full details about the license, please check the `LICENSE.md` file.
