--- packages/ckeditor5-dev-env/lib/tasks/create-package/templates/gulpfile.js
@@ -1,18 +1,0 @@
-/**
- * @license Copyright (c) 2003-2017, CKSource - Frederico Knabben. All rights reserved.
- * For licensing, see LICENSE.md.
- */
-
-/* jshint browser: false, node: true, strict: true */
-
-'use strict';
-
-const gulp = require( 'gulp' );
-
-const ckeditor5Lint = require( '@ckeditor/ckeditor5-dev-lint' )( {
-	ignoredFiles: 'src/lib/**'
-} );
-
-gulp.task( 'lint', ckeditor5Lint.lint );
-gulp.task( 'lint-staged', ckeditor5Lint.lintStaged );
-gulp.task( 'pre-commit', [ 'lint-staged' ] );
