--- packages/ckeditor5-dev-env/lib/tasks/create-package/templates/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "",
-  "version": "",
-  "description": "",
-  "keywords": [],
-  "dependencies": {},
-  "devDependencies": {
-  },
-  "engines": {
-    "node": ">=6.0.0",
-    "npm": ">=3.0.0"
-  },
-  "author": "CKSource (http://cksource.com/)",
-  "license": "See LICENSE.md",
-  "homepage": "https://ckeditor5.github.io",
-  "bugs": "https://github.com/{{GitHubRepositoryPath}}/issues",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/{{GitHubRepositoryPath}}.git"
-  }
-}
