--- packages/ckeditor5-dev-env/lib/tasks/exec/functions/git-push.js
@@ -1,21 +1,0 @@
-/**
- * @license Copyright (c) 2003-2017, CKSource - Frederico Knabben. All rights reserved.
- * For licensing, see LICENSE.md.
- */
-
-'use strict';
-
-const { git } = require( '@ckeditor/ckeditor5-dev-utils' );
-
-/**
- * Pushes changes of current branch in repository to default origin.
- *
- * Example:
- *
- *		gulp exec --task git-push
- *
- * @param {String} workdir
- */
-module.exports = function executeGitPush( workdir ) {
-	git.push( workdir );
-};
