+++ packages/ckeditor5-dev-env/tests/fixtures/basic/package.json
