+++ packages/ckeditor5-dev-env/tests/fixtures/basic/packages/ckeditor5-core/package.json
